LPS-106397 Move expression steps into step package.
