+++ modules/core/petra/petra-sql-dsl-api/src/main/resources/com/liferay/petra/sql/dsl/expression/step/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
