+++ modules/core/petra/petra-sql-dsl-spi/src/main/java/com/liferay/petra/sql/dsl/spi/expression/step/WhenThen.java
@@ -12,11 +12,12 @@
  * details.
  */
 
-package com.liferay.petra.sql.dsl.spi.expression;
+package com.liferay.petra.sql.dsl.spi.expression.step;
 
 import com.liferay.petra.sql.dsl.ast.ASTNodeListener;
-import com.liferay.petra.sql.dsl.expression.ElseEndStep;
 import com.liferay.petra.sql.dsl.expression.Expression;
+import com.liferay.petra.sql.dsl.expression.Predicate;
+import com.liferay.petra.sql.dsl.expression.step.WhenThenStep;
 import com.liferay.petra.sql.dsl.spi.ast.BaseASTNode;
 
 import java.util.Objects;
@@ -25,29 +26,40 @@ import java.util.function.Consumer;
 /**
  * @author Preston Crary
  */
-public class ElseEnd<T> extends BaseASTNode implements DefaultExpression<T> {
+public class WhenThen<T> extends BaseASTNode implements DefaultWhenThenStep<T> {
 
-	public ElseEnd(ElseEndStep<T> elseEndStep, Expression<T> elseExpression) {
-		super(elseEndStep);
+	public WhenThen(
+		WhenThenStep<T> whenThenStep, Predicate predicate,
+		Expression<T> thenExpression) {
 
-		_elseExpression = Objects.requireNonNull(elseExpression);
+		super(whenThenStep);
+
+		_predicate = Objects.requireNonNull(predicate);
+		_thenExpression = Objects.requireNonNull(thenExpression);
 	}
 
-	public Expression<T> getElseExpression() {
-		return _elseExpression;
+	public Predicate getPredicate() {
+		return _predicate;
+	}
+
+	public Expression<T> getThenExpression() {
+		return _thenExpression;
 	}
 
 	@Override
 	protected void doToSQL(
 		Consumer<String> consumer, ASTNodeListener astNodeListener) {
 
-		consumer.accept("else ");
+		consumer.accept("when ");
 
-		_elseExpression.toSQL(consumer, astNodeListener);
+		_predicate.toSQL(consumer, astNodeListener);
 
-		consumer.accept(" end");
+		consumer.accept(" then ");
+
+		_thenExpression.toSQL(consumer, astNodeListener);
 	}
 
-	private final Expression<T> _elseExpression;
+	private final Predicate _predicate;
+	private final Expression<T> _thenExpression;
 
 }
