+++ modules/core/petra/petra-sql-dsl-spi/src/main/resources/com/liferay/petra/sql/dsl/spi/expression/step/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
