+++ modules/core/petra/petra-sql-dsl-spi/src/test/java/com/liferay/petra/sql/dsl/spi/SQLDSLTest.java
@@ -22,7 +22,7 @@ import com.liferay.petra.sql.dsl.ast.ASTNodeListener;
 import com.liferay.petra.sql.dsl.expression.Alias;
 import com.liferay.petra.sql.dsl.expression.Expression;
 import com.liferay.petra.sql.dsl.expression.Predicate;
-import com.liferay.petra.sql.dsl.expression.WhenThenStep;
+import com.liferay.petra.sql.dsl.expression.step.WhenThenStep;
 import com.liferay.petra.sql.dsl.query.DSLQuery;
 import com.liferay.petra.sql.dsl.query.FromStep;
 import com.liferay.petra.sql.dsl.query.GroupByStep;
@@ -36,16 +36,16 @@ import com.liferay.petra.sql.dsl.spi.ast.BaseASTNode;
 import com.liferay.petra.sql.dsl.spi.ast.DefaultASTNodeListener;
 import com.liferay.petra.sql.dsl.spi.expression.AggregateExpression;
 import com.liferay.petra.sql.dsl.spi.expression.AliasImpl;
-import com.liferay.petra.sql.dsl.spi.expression.CaseWhenThen;
 import com.liferay.petra.sql.dsl.spi.expression.DSLFunction;
 import com.liferay.petra.sql.dsl.spi.expression.DSLFunctionType;
-import com.liferay.petra.sql.dsl.spi.expression.ElseEnd;
 import com.liferay.petra.sql.dsl.spi.expression.NullExpression;
 import com.liferay.petra.sql.dsl.spi.expression.Operand;
 import com.liferay.petra.sql.dsl.spi.expression.PredicateImpl;
 import com.liferay.petra.sql.dsl.spi.expression.Scalar;
 import com.liferay.petra.sql.dsl.spi.expression.ScalarList;
-import com.liferay.petra.sql.dsl.spi.expression.WhenThen;
+import com.liferay.petra.sql.dsl.spi.expression.step.CaseWhenThen;
+import com.liferay.petra.sql.dsl.spi.expression.step.ElseEnd;
+import com.liferay.petra.sql.dsl.spi.expression.step.WhenThen;
 import com.liferay.petra.sql.dsl.spi.query.From;
 import com.liferay.petra.sql.dsl.spi.query.GroupBy;
 import com.liferay.petra.sql.dsl.spi.query.Having;
